requirements.txt
langchain
langchain-community
faiss-cpu
python-dotenv
streamlit
langchain-google-genai
google-generativeai